requirements.txt
yt-dlp>=2023.12.30
smbprotocol>=1.12.0
requests>=2.31.0
python-dotenv>=1.0.0
